refactor: Rag Answer 분리
